+++ sample-apps/react/egress-composite/src/main.tsx
@@ -1,7 +1,7 @@
 import { createRoot } from 'react-dom/client';
 import { CompositeApp } from './CompositeApp';
 
-import '@stream-io/video-styling/dist/css/styles.css';
+import '@stream-io/video-react-sdk/dist/css/styles.css';
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <CompositeApp />,