+++ sample-apps/react/livestream-app/src/viewers/HLSLivestream.tsx
@@ -2,8 +2,7 @@ import './HLSLivestream.scss';
 
 import {
   LoadingIndicator,
-  useCallMetadata,
-  useIsCallBroadcastingInProgress,
+  useCallStateHooks,
 } from '@stream-io/video-react-sdk';
 import { useEffect, useMemo, useState } from 'react';
 import HLS from 'hls.js';
@@ -12,6 +11,8 @@ import { ViewerControls } from './ui/ViewerControls';
 import { Lobby } from './ui/Lobby';
 
 export const HLSLivestreamUI = () => {
+  const { useIsCallBroadcastingInProgress, useCallMetadata } =
+    useCallStateHooks();
   const isBroadcasting = useIsCallBroadcastingInProgress();
   const metadata = useCallMetadata();
   const hls = useMemo(() => new HLS(), []);